+++ Src/NeoIVClock/components/i2c_wrapper/include/i2c_wrapper.h
@@ -12,7 +12,7 @@ void i2c_wrapper_init(void);
 
 void i2c_wrapper_add_dev(
     uint16_t addr, 
-    uint16_t clk_speed_hz,
+    uint32_t clk_speed_hz,
     i2c_wrapper_dev_handle_t * dev_handle);
 
 typedef enum _i2c_wrapper_addr_mode_t
